+++ backend/app/builders/ProcessoBuilder.java
@@ -548,6 +548,14 @@ public class ProcessoBuilder extends CriteriaBuilder<Processo> {
 		return this;
 	}
 
+	public ProcessoBuilder groupByIdAnaliseTecnica(){
+
+		addAnaliseTecnicaAlias(true);
+		addProjection(Projections.groupProperty(ANALISE_TECNICA_ALIAS+".id").as("idAnaliseTecnica"));
+
+		return this;
+	}
+
 
 	public ProcessoBuilder groupByDataVencimentoPrazoAnaliseTecnica(boolean isLeftOuterJoin) {