+++ backend/app/deserializers/DateDeserializer.java
@@ -14,6 +14,7 @@ import java.util.Date;
 public class DateDeserializer implements JsonDeserializer<Date> {
 
     public static final String DATE_FORMAT = Play.configuration.getProperty("date.format");
+    private static final String DATE_FORMAT_TIMEABLE = Play.configuration.getProperty("date.format.timeable");
 
     @Override
     public Date deserialize(JsonElement json, Type type, JsonDeserializationContext context) throws JsonParseException {
@@ -22,7 +23,7 @@ public class DateDeserializer implements JsonDeserializer<Date> {
 
     public static Date parseDate(String dateText) {
         try {
-          return new SimpleDateFormat(DATE_FORMAT).parse(dateText);
+          return new SimpleDateFormat(DATE_FORMAT_TIMEABLE).parse(dateText);
         } catch (ParseException e) {
           e.printStackTrace();
           return null;