+++ frontend/src/features/analiseTecnica/analise-tecnica.controller.js
@@ -34,11 +34,11 @@ var AnaliseTecnicaController = function ($rootScope, uploadService, $route, $sco
     ctrl.tipoLicenca =  {};
     ctrl.inconsistenciasAdicionadas = [];
     ctrl.notificacao = {};
-	ctrl.notificacao.documentacao = null;
-	ctrl.notificacao.retificacaoEmpreendimento = null;
-	ctrl.notificacao.retificacaoSolicitacao = null;
-	ctrl.notificacao.retificacaoSolicitacaoComGeo = null;
-	ctrl.notificacao.prazoNotificacao = null;
+    ctrl.notificacao.documentacao = null;
+    ctrl.notificacao.retificacaoEmpreendimento = null;
+    ctrl.notificacao.retificacaoSolicitacao = null;
+    ctrl.notificacao.retificacaoSolicitacaoComGeo = null;
+    ctrl.notificacao.prazoNotificacao = null;
 
     ctrl.parecer = {
         doProcesso: null,
@@ -1392,7 +1392,7 @@ var AnaliseTecnicaController = function ($rootScope, uploadService, $route, $sco
 
         }
 
-		if(!camposConclusaoValidos()) {
+        if(!camposConclusaoValidos()) {
 
             mensagem.error('Não foi possível concluir a análise. Verifique os campos obrigatórios!', { ttl: 10000 });
             return;
@@ -1466,7 +1466,18 @@ var AnaliseTecnicaController = function ($rootScope, uploadService, $route, $sco
                                         a.click();
 
                                         if(ctrl.parecer.vistoria.realizada) {
-                                            //RTV
+                                            
+                                            documentoAnaliseService.generatePDFRelatorioTecnicoVistoria(params)
+                                                .then(function(data) {
+
+                                                        var a = document.createElement('a');
+                                                        a.href = URL.createObjectURL(data.data.response.blob);
+                                                        a.download = data.data.response.fileName ? data.data.response.fileName : 'relatorio_tecnico_vistoria.pdf';
+                                                        a.click();
+
+                                                },function(error){
+                                                    mensagem.error(error.data.texto);
+                                            });
                                         }
 
                                     }, function(error){
@@ -1474,7 +1485,18 @@ var AnaliseTecnicaController = function ($rootScope, uploadService, $route, $sco
                                     });
 
                             } else if(ctrl.parecer.vistoria.realizada) {
-                                // RTV
+                                
+                                documentoAnaliseService.generatePDFRelatorioTecnicoVistoria(params)
+                                    .then(function(data) {
+
+                                            var a = document.createElement('a');
+                                            a.href = URL.createObjectURL(data.data.response.blob);
+                                            a.download = data.data.response.fileName ? data.data.response.fileName : 'relatorio_tecnico_vistoria.pdf';
+                                            a.click();
+
+                                    },function(error){
+                                        mensagem.error(error.data.texto);
+                                });
                             }
 
                         }, function(error){
@@ -1482,7 +1504,18 @@ var AnaliseTecnicaController = function ($rootScope, uploadService, $route, $sco
                         });
 
                 } else if(ctrl.parecer.vistoria.realizada) {
-                    //RTV 
+
+                    documentoAnaliseService.generatePDFRelatorioTecnicoVistoria(params)
+                        .then(function(data) {
+
+                                var a = document.createElement('a');
+                                a.href = URL.createObjectURL(data.data.response.blob);
+                                a.download = data.data.response.fileName ? data.data.response.fileName : 'relatorio_tecnico_vistoria.pdf';
+                                a.click();
+
+                        },function(error){
+                            mensagem.error(error.data.texto);
+                    }); 
                 }
 
                 $location.path('/analise-tecnica');