+++ frontend/src/services/documentoAnalise.service.js
@@ -46,6 +46,11 @@ var DocumentoAnaliseService = function(request, config, $window) {
 		return request.postArrayBuffer(config.BASE_URL() + 'analisesTecnicas/downloadPDFMinuta', analiseTecnica);
 	};
 
+	this.generatePDFRelatorioTecnicoVistoria = function(analiseTecnica) {
+
+		return request.postArrayBuffer(config.BASE_URL() + 'analisesTecnicas/downloadPDFRelatorioTecnicoVistoria', analiseTecnica);
+	};
+
 };
 
 exports.services.DocumentoAnaliseService = DocumentoAnaliseService;