+++ CHANGELOG.md
@@ -48,3 +48,4 @@
 * OS35 - US945 - Eu, analista técnico, devo finalizar análise para gerar documento pdf do parecer
 * OS35 - US304 - Eu, analista técnico, devo acessar o menu "Consulta protocolo" para consultar os processos
 * OS35 - US295 - Eu, analista técnico, devo visualizar o protocolo para analisar suas informações
+* OS35 - US958 - Eu, analista técnico, devo finalizar análise para gerar documento pdf do Relatório Técnico de Vistoria